third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/gen-reconnect-during.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let service;
  return getHealthThermometerService()
    .then(vals => ({service} = vals))
    .then(() => service.getCharacteristic('heart_rate_measurement')
    .then(() => assert_unreached('Got descriptor unexpectedly.'), (e) => {
      assert_equals(e.name, 'NetworkError');
      assert_equals(e.message,
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.'));
    });
}, 'disconnect() and connect() called during getCharacteristic. Reject with ' +
   'NetworkError.');

</script>
